requirements.txt
alembic==1.14.1
bcrypt==4.0.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==41.0.5
dnspython==2.7.0
email-validator==2.1.0.post1
exceptiongroup==1.2.2
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
gunicorn==21.2.0
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
Pillow==10.0.1
pluggy==1.5.0
pycparser==2.22
PyMySQL==1.1.0
pypng==0.20220715.0
pytest==7.4.3
pytest-flask==1.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
qrcode==7.4.2
requests==2.31.0
six==1.17.0
SQLAlchemy==2.0.38
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==2.3.7
WTForms==3.2.1
zipp==3.21.0
